% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_two.R
\name{mean_and_two}
\alias{mean_and_two}
\title{mean_and_med}
\usage{
mean_and_two(vec)
}
\arguments{
\item{vec}{A numeric vector}
}
\value{
The mean of this vector plus two, ignoring the NAs
}
\description{
mean_and_med
}
\examples{

mean_and_two(vec = c(10, 14, 13))
}
